Modified

+++ students-api/src/main/java/com/martin/studentsapi/studentsapi/services/impl/StudentServiceImpl.java
@@ -2,7 +2,12 @@ package com.martin.studentsapi.studentsapi.services.impl;
 
 import com.martin.studentsapi.studentsapi.models.Student;
 import com.martin.studentsapi.studentsapi.models.StudyProgram;
+import com.martin.studentsapi.studentsapi.models.exceptions.IndexException;
+import com.martin.studentsapi.studentsapi.models.exceptions.MissingParameterException;
+import com.martin.studentsapi.studentsapi.models.exceptions.StudentNotFoundException;
+import com.martin.studentsapi.studentsapi.models.exceptions.StudyProgramNotFoundException;
 import com.martin.studentsapi.studentsapi.repositories.StudentRepository;
+import com.martin.studentsapi.studentsapi.repositories.StudyProgramRepository;
 import com.martin.studentsapi.studentsapi.services.StudentService;
 import org.springframework.stereotype.Service;
 
@@ -17,9 +22,11 @@ import java.util.stream.Collectors;
 public class StudentServiceImpl implements StudentService {
 
     private final StudentRepository studentRepository;
+    private final StudyProgramRepository studyProgramRepository;
 
-    public StudentServiceImpl(StudentRepository studentRepository) {
+    public StudentServiceImpl(StudentRepository studentRepository, StudyProgramRepository studyProgramRepository) {
         this.studentRepository = studentRepository;
+        this.studyProgramRepository = studyProgramRepository;
     }
 
     @Override
@@ -28,12 +35,30 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student addNew(String index, String name, String lastName) {
+    public Student addNew(String index, String name, String lastName, String studyProgramName) throws MissingParameterException, IndexException, StudyProgramNotFoundException {
         Student student = new Student();
+
+        if(index == null || name == null || lastName == null || studyProgramName == null)
+            throw new MissingParameterException();
+
+        if(index.length() != 6)
+            throw new IndexException();
+
+        List<StudyProgram> studyPrograms = (List<StudyProgram>) studyProgramRepository.findAll();
+
+        Optional<StudyProgram> studyProgram = studyPrograms
+                .stream()
+                .filter(s -> s.getName().equals(studyProgramName))
+                .findAny();
+
+        if(!studyProgram.isPresent())
+            throw new StudyProgramNotFoundException();
+
+
         student.setIndex(index);
         student.setName(name);
         student.setLastName(lastName);
-//        student.setStudyProgram(new StudyProgram(studyProgramName));
+        student.setStudyProgram(studyProgram.get());
 
         studentRepository.save(student);
 
@@ -41,13 +66,16 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student deleteStudent(String index) {
+    public Student deleteStudent(String index) throws StudentNotFoundException {
         List<Student> students = (List<Student>) studentRepository.findAll();
 
         Optional<Student> student = students
                 .stream()
                 .filter(s -> s.getIndex().equals(index))
                 .findFirst();
+
+        if(!student.isPresent())
+            throw new StudentNotFoundException();
 
         studentRepository.delete(student.get());
 
@@ -67,12 +95,26 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student updateStudent(String index, String name, String lastName) {
+    public Student updateStudent(String index, String name, String lastName, String studyProgramName) throws StudentNotFoundException {
 
         Optional<Student> student = studentRepository.findById(index);
 
+        if(!student.isPresent())
+            throw new StudentNotFoundException();
+
         student.get().setName(name);
         student.get().setLastName(lastName);
+
+        List<StudyProgram> studyPrograms = (List<StudyProgram>) studyProgramRepository.findAll();
+
+        Optional<StudyProgram> studyProgram = studyPrograms.stream().filter(s -> s.getName().equals(studyProgramName)).findFirst();
+
+        if(studyProgram.isPresent()){
+            student.get().setStudyProgram(studyProgram.get());
+        }
+        else {
+            throw new StudentNotFoundException();
+        }
 
         studentRepository.save(student.get());
